WEB-4526 extract agpUtils into abstracted function
